requirements.txt
selenium~=4.29.0
allure-python-commons~=2.13.5
pytest~=8.3.4